api/requirements.txt
fastapi==0.110.3
uvicorn[standard]==0.29.0
python-multipart==0.0.9

pydantic==2.7.4

numpy==1.26.4
opencv-python-headless==4.10.0.84
Pillow==10.4.0
pymupdf
openpyxl==3.1.5
python-docx==1.1.2
opencv-python
reportlab==4.2.2
supabase==2.6.0

pandas==2.2.2
xlsxwriter==3.2.0
python-dotenv==1.0.1
requests==2.32.3
pymupdf==1.24.10